+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/configuration/ContextBrokerConfig.java
@@ -22,6 +22,7 @@ public class ContextBrokerConfig {
             subscriptionService.createDefaultSubscription();
         } catch (Exception e) {
             log.error("Error creating default subscription", e);
+            Thread.currentThread().interrupt();
         }
 
     }