+++ src/main/java/es/in2/dome/blockchain/connector/utils/ApplicationUtils.java
@@ -2,7 +2,6 @@ package es.in2.dome.blockchain.connector.utils;
 
 import es.in2.dome.blockchain.connector.exception.RequestErrorException;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 
 import java.net.URI;
@@ -10,9 +9,7 @@ import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
 
 @Slf4j
 @Component
@@ -42,7 +39,7 @@ public class ApplicationUtils {
         String statusCode = response.thenApply(HttpResponse::statusCode).join().toString();
         String body = response.thenApply(HttpResponse::body).join();
 
-        if(statusCode.equals("200")) {
+        if (statusCode.equals("200")) {
             log.debug("Request successful");
         } else if (statusCode.equals("404")) {
             log.error("Not found");
@@ -71,15 +68,34 @@ public class ApplicationUtils {
         checkPostResponse(response);
     }
 
+    public void patchRequest(String url, String requestBody) {
+        // Create request
+        HttpClient client = HttpClient.newHttpClient();
+        HttpRequest request = HttpRequest.newBuilder()
+                .uri(URI.create(url))
+                .headers("Content-Type", "application/json")
+                .method("PATCH", HttpRequest.BodyPublishers.ofString(requestBody))  // Use PATCH method
+                .build();
+
+        // Send request asynchronously
+        CompletableFuture<HttpResponse<String>> response =
+                client.sendAsync(request, HttpResponse.BodyHandlers.ofString());
+
+        // Verify Response HttpStatus
+        checkPostResponse(response);
+    }
+
     private void checkPostResponse(CompletableFuture<HttpResponse<String>> response) {
         String statusCode = response.thenApply(HttpResponse::statusCode).join().toString();
         String body = response.thenApply(HttpResponse::body).join();
-        if(statusCode.equals("201")) {
-            log.debug("Request successful");
-        } else {
-            log.error("Bad Request");
-            throw new RequestErrorException("Bad Request: " + body);
+
+        switch (statusCode) {
+            case "201", "200" -> log.debug("Request successful");
+            case "204" -> log.debug("Successful Patch");
+            default -> {
+                log.error("Bad Request");
+                throw new RequestErrorException("Bad Request: " + body);
+            }
         }
     }
-
 }